+++ esp_prototype/main/data_center/Account.cc
@@ -27,7 +27,7 @@ Account::Account(const char* id, DataCenter* center, uint32_t bufSize, void* use
         uint8_t* buf1 = buffer + bufSize;
 
         _priv.bufferManager.init(buf0, buf1);
-        DC_LOG_INFO("Account[%s] cached %d x2 bytes", _id, bufSize);
+        DC_LOG_INFO("Account[%s] cached %lu x2 bytes", _id, static_cast<unsigned long>(bufSize));
         _priv.bufferSize = bufSize;
     }
 
@@ -113,8 +113,8 @@ bool Account::commit(const void* data_p, uint32_t size) {
 
     _priv.bufferManager.setWriteDone();
 
-    DC_LOG_INFO("pub[%s] commit data(0x%p)[%d] >> data(0x%p)[%d] done",
-                _id, data_p, size, wBuf, size);
+    DC_LOG_INFO("pub[%s] commit data(0x%p)[%lu] >> data(0x%p)[%lu] done",
+                _id, data_p, static_cast<unsigned long>(size), wBuf, static_cast<unsigned long>(size));
 
     return true;
 }
@@ -143,8 +143,8 @@ int Account::publish() {
     for (auto sub : _subscribers) {
         EventCallback callback = sub->_priv.eventCallback;
 
-        DC_LOG_INFO("pub[%s] publish >> data(0x%p)[%d] >> sub[%s]...",
-                    _id, param.data_p, param.size, sub->_id);
+        DC_LOG_INFO("pub[%s] publish >> data(0x%p)[%lu] >> sub[%s]...",
+                    _id, param.data_p, static_cast<unsigned long>(param.size), sub->_id);
 
         if (callback != nullptr) {
             param.recv = sub;
@@ -179,8 +179,8 @@ int Account::pull(Account* pub, void* data_p, uint32_t size) {
         return RES_NOT_FOUND;
     }
 
-    DC_LOG_INFO("sub[%s] pull << data(0x%p)[%d] << pub[%s] ...",
-                _id, data_p, size, pub->_id);
+    DC_LOG_INFO("sub[%s] pull << data(0x%p)[%lu] << pub[%s] ...",
+                _id, data_p, static_cast<unsigned long>(size), pub->_id);
 
     EventCallback callback = pub->_priv.eventCallback;
     if (callback != nullptr) {
@@ -211,8 +211,8 @@ int Account::pull(Account* pub, void* data_p, uint32_t size) {
                 DC_LOG_WARN("pub[%s] data was not commit!", pub->_id);
             }
         } else {
-            DC_LOG_ERROR("Data size pub[%s]:%d != sub[%s]:%d",
-                         pub->_id, pub->_priv.bufferSize, _id, size);
+            DC_LOG_ERROR("Data size pub[%s]:%lu != sub[%s]:%lu",
+                         pub->_id, static_cast<unsigned long>(pub->_priv.bufferSize), _id, static_cast<unsigned long>(size));
         }
     }
 
@@ -235,8 +235,8 @@ int Account::notify(Account* pub, const void* data_p, uint32_t size) {
         return RES_NOT_FOUND;
     }
 
-    DC_LOG_INFO("sub[%s] notify >> data(0x%p)[%d] >> pub[%s] ...",
-                _id, data_p, size, pub->_id);
+    DC_LOG_INFO("sub[%s] notify >> data(0x%p)[%lu] >> pub[%s] ...",
+                _id, data_p, static_cast<unsigned long>(size), pub->_id);
 
     EventCallback callback = pub->_priv.eventCallback;
     if (callback != nullptr) {
